Move Variant to its own file, cleanup.
