+++ Libs/Project/Settings.h
@@ -2,101 +2,10 @@
 
 #include <map>
 #include <string>
-#include <vector>
-#include <sstream>
-#include <iterator>
-#include <algorithm>
-#include <iostream>
-#include <memory>
 
-namespace shapeworks
-{
+#include "Variant.h"
 
-// class for automatic conversion from string -> anytype. (adapted from optparse)
-class Variant
-{
-public:
-
-  Variant() : str(), valid(false) {}
-
-  Variant(const std::string &v) : str(v), valid(true) {}
-  Variant(int v) : str(std::to_string(v)), valid(true) {}
-  Variant(double v) : str(std::to_string(v)), valid(true) {}
-  Variant(const char* v) : str(v), valid(true) {}
-  Variant(bool v) : str(v ? "true" : "false"), valid(true) {}
-
-  operator std::vector<std::string>() {
-    std::istringstream iss(str);
-    std::vector<std::string> v((std::istream_iterator<std::string>(iss)),
-                               std::istream_iterator<std::string>());
-    return v;
-  }
-
-  operator std::string() {
-    return str;
-  }
-
-  operator bool() {
-    // first try to read as the word 'true' or 'false', otherwise 0 or 1
-    std::string str_lower(str);
-    std::transform(str.begin(), str.end(), str_lower.begin(),
-                   [](unsigned char c) { return std::tolower(c); });
-    bool t = (valid && (std::istringstream(str_lower) >> std::boolalpha >> t)) ? t : false;
-    if (!t) {
-      t = (valid && (std::istringstream(str) >> t)) ? t : false;
-    }
-    return t;
-  }
-
-  operator short() {
-    short t;
-    return (valid && (std::istringstream(str) >> t)) ? t : 0;
-  }
-
-  operator unsigned short() {
-    unsigned short t;
-    return (valid && (std::istringstream(str) >> t)) ? t : 0;
-  }
-
-  operator int() {
-    int t;
-    return (valid && (std::istringstream(str) >> t)) ? t : 0;
-  }
-
-  operator unsigned int() {
-    unsigned int t;
-    return (valid && (std::istringstream(str) >> t)) ? t : 0;
-  }
-
-  operator long() {
-    long t;
-    return (valid && (std::istringstream(str) >> t)) ? t : 0;
-  }
-
-  operator unsigned long() {
-    unsigned long t;
-    return (valid && (std::istringstream(str) >> t)) ? t : 0;
-  }
-
-  operator float() {
-    float t;
-    return (valid && (std::istringstream(str) >> t)) ? t : 0;
-  }
-
-  operator double() {
-    double t;
-    return (valid && (std::istringstream(str) >> t)) ? t : 0;
-  }
-
-  operator long double() {
-    long double t;
-    return (valid && (std::istringstream(str) >> t)) ? t : 0;
-  }
-private:
-
-  const std::string str;
-  bool valid;
-};
+namespace shapeworks {
 
 class Settings {
 
@@ -106,7 +15,6 @@ public:
   static constexpr const char* GROOM_SETTINGS = "groom";
   static constexpr const char* OPTIMIZE_SETTINGS = "optimize";
   static constexpr const char* STUDIO_SETTINGS = "studio";
-
 
   void set_map(std::map<std::string, std::string> map);
   std::map<std::string, std::string> get_map();
